requirements.txt
numpy>=1.21.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.1
opencv-python>=4.5.3
matplotlib>=3.4.3
scipy>=1.7.1
tqdm>=4.62.2

shapely>=1.8.0
PyYAML>=5.4.1

torchvision>=0.10.0
torch>=1.9.0

typing-extensions>=3.10.0

attrdict

lightning 